backend/requirements.txt
fastapi[standard]
google-genai
python-dotenv

